openssl: Fix Android detection in configuration step
Explicitly pass -D__ANDROID_API__=N to the OpenSSL configuration script.
This is needed for the script to infer some configuration values, as per
the ANDROID.NOTES file included in the source tree. Also, with newer
version of the NDK, the same file recommends *against* defining the
CROSS_SYSROOT variable, because Clang-based NDKs arrange themselves to
use the sysroot and setting it accidentally to values that does not match
the current architecture+API target results in either miscompilation or
failed build configuration.