CI: Run builds when targeting stable branches
Allow running the "build" job when targeting stable branches. If it was
possible to use a regular expression to match branch names, this would
use

  ^wpewebkit-[0-9]+\.[0-9]+$

but as that is not possible, this matches instead branches with the
"wpewebkit-2." prefix followed by two digits.

(cherry picked from commit d19b38d0e8afd754465c967026afbff5fcc002c1)